use utils.split for tracked-controls
